backend/pyproject.toml
[project]
name = "styleme-backend"
version = "0.1.0"
description = "StyleMe — AI-powered personal wardrobe stylist backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clip",
    "fastapi>=0.135.3",
    "google-genai>=1.70.0",
    "hydra-db-python>=0.1.5",
    "onnxruntime>=1.24.4",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=11.0.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "timm>=1.0.26",
    "ultralytics>=8.4.33",
    "uvicorn[standard]>=0.43.0",
]

[project.optional-dependencies]
face = [
    "insightface>=0.7.3",
]

[tool.uv.sources]
clip = { git = "https://github.com/ultralytics/CLIP.git" }
